+++ node_helper.js
@@ -7,7 +7,7 @@
 /* jshint node: true, esversion: 6*/
 
 var NodeHelper = require("node_helper");
-var newConsole = require("tracer").console({
+var tracer = require("tracer").console({
     format: "{{timestamp}} <{{title}}> {{message}} ({{folder}}/{{file}}:{{line}} {{method}})",
     overwriteConsoleMethods: true,
     dateformat: "yyyy-mm-dd'T'HH:MM:ss",
@@ -23,12 +23,18 @@ module.exports = NodeHelper.create({
         if (notification === "INITIALIZE_LOGGING" && !this.initialized) {
             this.config = payload;
             if (this.config.useColor) {
-                newConsole = require('tracer').colorConsole(this.config);
+                tracer = require('tracer').colorConsole(this.config);
             } else {
-                newConsole = require('tracer').console(this.config);
+                tracer = require('tracer').console(this.config);
             }
             console.info("MMM-Logging updated config received, reloading console");
             this.initialized = true;
+        }
+        if (notification === "NOTIFICATION_TO_CONSOLE") {
+            tracer.log("Module Notification: " + payload.notification + ((payload.sender) ? " from " + payload.sender : "") + (payload.payload) ? " payload: " + JSON.stringify(payload.payload, undefined, 3) : "");
+        }
+        if (notification === "BROWSER_ERROR") {
+            tracer.log("Browser Error: " + JSON.stringify(payload, undefined, 3));
         }
     },
 });